+++ lib/concierge/suppliers/rentals_united/entities/property.rb
@@ -8,7 +8,9 @@ module RentalsUnited
                     :postal_code, :check_in_time, :check_out_time, :max_guests,
                     :surface, :bedroom_type_id, :property_type_id, :floor,
                     :images, :amenities, :owner_id, :security_deposit_type,
-                    :security_deposit_amount, :check_in_instructions
+                    :security_deposit_amount, :check_in_instructions,
+                    :number_of_bathrooms, :number_of_single_beds,
+                    :number_of_double_beds, :number_of_sofa_beds
 
       attr_writer :active, :archived
 
@@ -17,7 +19,9 @@ module RentalsUnited
                      max_guests:, surface:, bedroom_type_id:, property_type_id:,
                      floor:, images:, amenities:, active:, archived:, owner_id:,
                      security_deposit_type:, security_deposit_amount:,
-                     check_in_instructions:)
+                     check_in_instructions:, number_of_bathrooms:,
+                     number_of_single_beds:, number_of_double_beds:,
+                     number_of_sofa_beds:)
         @id                      = id
         @title                   = title
         @description             = description
@@ -40,6 +44,10 @@ module RentalsUnited
         @archived                = archived
         @images                  = images
         @amenities               = amenities
+        @number_of_bathrooms     = number_of_bathrooms
+        @number_of_single_beds   = number_of_single_beds
+        @number_of_double_beds   = number_of_double_beds
+        @number_of_sofa_beds     = number_of_sofa_beds
       end
 
       def active?