+++ CHANGELOG.md
@@ -4,6 +4,15 @@ This file summarises the most important changes that went live on each release
 of Concierge. Please check the Wiki entry on the release process to understand
 how this file is formatted and how the process works.
 
+## [0.12.8] - 2016-10-17
+### Added
+- Rentals United: add number_of_bathrooms, number single, double, sofa beds fields
+- Synchronisation#mark_as_processed
+
+### Fixed
+- Rentals United: fix exception which was raised while parsing multi-lang check-in instructions
+- Kigo and Kigo Legacy metadata sync should not purge properties when rate limitted
+
 ## [0.12.7] - 2016-10-17
 ### Fixed
 - Ensure synchronisation.finish! for Kigo host with empty properties